Keep displaying contacts since association to JWT
This commit keeps displaying the contacts data almost the same way they
were presented when related to the user account (now the JWT).
Meaning the are still displayed in the "Contacts" tab of the dashboard,
only being regrouped by JWT now, especially the usage policy of the
token so it's easy to identify wich contact is related to which token.

The method `setContacts` in the store reformating the contact's data
received from the API is pretty ugly (in order to save headaches in
component's templates). But this solution has the benefits to quickly
maintain the ascending compatibility since we don't need a complete
rework (coming soon) of the interface to display the contacts after the
changes in the model.

Closes #68